Reorganize pt_BR company tests
